AnimationControl/EXECommandQueryCreate.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace AnimationControl
{
    public class EXECommandQueryCreate : EXECommand
    {
        private String ReferencingVariableName { get; }
        private String ClassName { get; }

        public EXECommandQueryCreate(String ClassName, String ReferencingVariableName)
        {
            this.ReferencingVariableName = ReferencingVariableName;
            this.ClassName = ClassName;
        }

        public EXECommandQueryCreate(String ClassName)
        {
            this.ReferencingVariableName = "";
            this.ClassName = ClassName;
        }

        // SetUloh2
        public override bool Execute(Animation Animation, EXEScope Scope)
        {
            //Create an instance of given class -> will affect ExecutionSpace.
            //If ReferencingVariableName is provided (is not ""), create a referencing variable pointing to this instance -> will affect scope
            CDClass Class = Animation.ExecutionSpace.getClassByName(this.ClassName);
            if (Class == null)
            {
                return false;
            }

            if (!"".Equals(this.ReferencingVariableName) && Scope.VariableNameExists(this.ReferencingVariableName))
            {
                return false;
            }

            CDClassInstance Instance = Class.CreateClassInstance();
            if (Instance == null)
            {
                return false;
            }

            if (!"".Equals(this.ReferencingVariableName))
            {
                EXEReferencingVariable Variable = new EXEReferencingVariable(this.ReferencingVariableName, Class.Name, Instance.UniqueID);
                return Scope.AddVariable(Variable);
            }

            return true;
        }
    }
}
